Address feedback and reorganize cli config code
