+++ typescript/cli/package.json
@@ -53,6 +53,5 @@
     "Deployment",
     "Typescript"
   ],
-  "packageManager": "yarn@3.2.0",
-  "stableVersion": "1.4.2"
+  "packageManager": "yarn@3.2.0"
 }
