+++ typescript/cli/src/config/warp.ts
@@ -1,19 +1,72 @@
 import { confirm, input } from '@inquirer/prompts';
 import { ethers } from 'ethers';
+import { z } from 'zod';
 
 import { TokenType } from '@hyperlane-xyz/hyperlane-token';
 
-import {
-  WarpRouteConfig,
-  isValidWarpRouteConfig,
-  readChainConfigIfExists,
-} from '../configs.js';
 import { errorRed, logBlue, logGreen } from '../logger.js';
 import {
   runMultiChainSelectionStep,
   runSingleChainSelectionStep,
 } from '../utils/chains.js';
-import { FileFormat, writeYamlOrJson } from '../utils/files.js';
+import { FileFormat, readYamlOrJson, writeYamlOrJson } from '../utils/files.js';
+
+import { readChainConfigIfExists } from './chain.js';
+
+const ConnectionConfigSchema = {
+  mailbox: z.string().optional(),
+  interchainGasPaymaster: z.string().optional(),
+  interchainSecurityModule: z.string().optional(),
+  foreignDeployment: z.string().optional(),
+};
+
+export const WarpRouteConfigSchema = z.object({
+  base: z.object({
+    type: z.literal(TokenType.native).or(z.literal(TokenType.collateral)),
+    chainName: z.string(),
+    address: z.string().optional(),
+    isNft: z.boolean().optional(),
+    name: z.string().optional(),
+    symbol: z.string().optional(),
+    decimals: z.number().optional(),
+    ...ConnectionConfigSchema,
+  }),
+  synthetics: z
+    .array(
+      z.object({
+        chainName: z.string(),
+        name: z.string().optional(),
+        symbol: z.string().optional(),
+        totalSupply: z.number().optional(),
+        ...ConnectionConfigSchema,
+      }),
+    )
+    .nonempty(),
+});
+
+type InferredType = z.infer<typeof WarpRouteConfigSchema>;
+// A workaround for Zod's terrible typing for nonEmpty arrays
+export type WarpRouteConfig = {
+  base: InferredType['base'];
+  synthetics: Array<InferredType['synthetics'][0]>;
+};
+
+export function readWarpRouteConfig(filePath: string) {
+  const config = readYamlOrJson(filePath);
+  if (!config) throw new Error(`No warp config found at ${filePath}`);
+  const result = WarpRouteConfigSchema.safeParse(config);
+  if (!result.success) {
+    const firstIssue = result.error.issues[0];
+    throw new Error(
+      `Invalid warp config: ${firstIssue.path} => ${firstIssue.message}`,
+    );
+  }
+  return result.data;
+}
+
+export function isValidWarpRouteConfig(config: any) {
+  return WarpRouteConfigSchema.safeParse(config).success;
+}
 
 export async function createWarpConfig({
   format,
